% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPBIB5.R
\name{GPBIB5}
\alias{GPBIB5}
\title{Generalized rectangular right angular (5) design.}
\usage{
GPBIB5(n, l, s, w)
}
\arguments{
\item{n}{Number of lines of the association schemes array.}

\item{l}{Number of columns of the association schemes array.}

\item{s}{Number of the token treatments from the same row of the association scheme.}

\item{w}{Number of the association scheme arrays.}
}
\value{
A LIST :
 \itemize{
  \item \code{PBIB } The configuration of the PBIB.
  \item \code{Type } The type of the design
  \item \code{V } Number of treatments.
  \item \code{B } Number of blocs.
  \item \code{R } Repetition of each treatment.
  \item \code{K } Size of blocs.
  \item \code{lamda } Vector of m-lambda.
  \item \code{Resolvable } Is the design Resolvable ?
  }
}
\description{
gives the configuration and the parametres of the design obtained
by the construction method of GPBIB_5 (see 3.2 of the paper rezgui
et al (2015)).
}
\note{
For \eqn{w=2}, the \code{GPBIB_5} is a rectangular right angular (5) (PBIB_5).
}
\examples{
\dontrun{
n<-3
l<-3
s<-3
w<-3
GPBIB5(n, l, s, w)
}
}
\references{
Imane Rezgui, Z. Gheribi-Aoulmi and H. Monod (2015). U-type Designs via
New Generalized Partially Balanced Incomplete Block Designs with m = 4, 5
and 7 Associated Classes, \doi{10.4236/am.2015.62024}, Applied mathematics, 6, 242-264.

Imane Rezgui, Z.Gheribi-Aoulmi and H. Monod, New association schemes
with 4, 5 and 7 associated classes and their associated partially
balanced incomplete block designs; Advances and Applications in Discrete
Mathematics Vol.12 Issue 2 197-206.
}
\seealso{
\code{\link{UType}}
}
\author{
Mohamed Laib, Imane Rezgui, Zebida Gheribi-Aoulmi and Herve Monod
}
