\name{CombinS-package}
\alias{CombinS-package}
\alias{CombinS}
\docType{package}
\title{
Constructions method of rectangular PBIB and rectangular right angular PBIB(m) (m= 4,5 and 7) designs
}
\description{
Methods of construction of rectangular PBIB designs from a  rectangular association scheme  and rectangular right angular PBIB(m) designs from rectangular right angular association scheme (m) (m= 4,5 and 7)
}
\details{
\tabular{ll}{
Package: \tab CombinS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-25\cr
License: \tab GPL-3\cr
}
}
\author{
Mohamed LAIB, Imane REZGUI, Zoubida Gheribi-Aoulmi and Herve Monod

Maintainer: Mohamed LAIB <laib.med@gmail.com>
}
\references{
Imane Rezgui, Z.Gheribi-Aoulmi and H. Monod, New association schemes with 4, 5 and 7 associated classes and their associated partially balanced incomplete block designs; Advances and Applications in Discrete Mathematics Vol.12 Issue 2 (to appear).
}
\keyword{ package }
\examples{
#*************************
#CombS
#***
SGDD<-CombS(4,3,3)
RD<-CombS(3,4,3)
}
