% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UType.R
\name{UType}
\alias{UType}
\title{U-type design via some PBIB designs}
\usage{
UType(lst)
}
\arguments{
\item{lst}{The output of one of our package functions.}
}
\value{
A LIST :
 \itemize{
  \item \code{v} Number of runs.
  \item \code{r } Number of factors.
  \item \code{UtypeDesign } The configuration of the U-type design..
  }
}
\description{
Applies the Fang algorithm on our constructed designs to obtain
the configuration and the parameters of the U-type design associated.
}
\examples{
\dontrun{
M<-GPBIB4A(4,4,2,2)
UType(M)
}
}
\references{
K.T. Fang, R.Li and A.Sudjanto (2006). Design ans Modeling for Computer
Experiments. Taylor & Francis Group, LLC London.
}
\author{
Mohamed Laib, Imane Rezgui, Zebida Gheribi-Aoulmi and Herve Monod
}
